Create Route and fix directions in folder
