+++ src/components/footer/footer.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import Logo from '../logo/logo';
+
+function Footer(): JSX.Element {
+  return (
+    <footer className="page-footer">
+      <Logo />
+      <div className="copyright">
+        <p>© 2019 What to watch Ltd.</p>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
