+++ src/components/logo/logo.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+function Logo(): JSX.Element {
+  return (
+    <div className="logo">
+      <Link to='/' className="logo__link logo__link--light">
+        <span className="logo__letter logo__letter--1">W</span>
+        <span className="logo__letter logo__letter--2">T</span>
+        <span className="logo__letter logo__letter--3">W</span>
+      </Link>
+    </div>
+  );
+}
+
+export default Logo;
