+++ src/components/movies-card/movies-card.tsx
