+++ src/components/player-pause/player-pause.tsx
