+++ src/components/player/player.tsx
