+++ src/components/sign-in-error/sign-in-error.tsx
@@ -1,29 +1,24 @@
 import React from 'react';
+import Footer from '../footer/footer';
+import Logo from '../logo/logo';
 
-function SignMessage(): JSX.Element {
+function SignError(): JSX.Element {
   return (
     <React.Fragment>
       <div className="visually-hidden"></div>
       <div className="user-page">
         <header className="page-header user-page__head">
-          <div className="logo">
-            <a href="main.html" className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
+          <Logo />
           <h1 className="page-title user-page__title">Sign in</h1>
         </header>
 
         <div className="sign-in user-page__content">
           <form action="#" className="sign-in__form">
             <div className="sign-in__message">
-              <p>We can’t recognize this email <br /> and password combination. Please try again.</p>
+              <p>Please enter a valid email address</p>
             </div>
             <div className="sign-in__fields">
-              <div className="sign-in__field">
+              <div className="sign-in__field sign-in__field--error">
                 <input className="sign-in__input" type="email" placeholder="Email address" name="user-email" id="user-email" />
                 <label className="sign-in__label visually-hidden" htmlFor="user-email">Email address</label>
               </div>
@@ -37,23 +32,10 @@ function SignMessage(): JSX.Element {
             </div>
           </form>
         </div>
-
-        <footer className="page-footer">
-          <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
-          <div className="copyright">
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
+        <Footer />
       </div>
     </React.Fragment>
   );
 }
 
-export default SignMessage;
+export default SignError;
