+++ src/components/sign-in-message/sign-in-message.tsx
@@ -1,24 +1,22 @@
 import React from 'react';
+import Footer from '../footer/footer';
+import Logo from '../logo/logo';
 
-function Sign(): JSX.Element {
+function SignMessage(): JSX.Element {
   return (
     <React.Fragment>
       <div className="visually-hidden"></div>
       <div className="user-page">
         <header className="page-header user-page__head">
-          <div className="logo">
-            <a href="main.html" className="logo__link">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
+          <Logo />
           <h1 className="page-title user-page__title">Sign in</h1>
         </header>
 
         <div className="sign-in user-page__content">
           <form action="#" className="sign-in__form">
+            <div className="sign-in__message">
+              <p>We can’t recognize this email <br /> and password combination. Please try again.</p>
+            </div>
             <div className="sign-in__fields">
               <div className="sign-in__field">
                 <input className="sign-in__input" type="email" placeholder="Email address" name="user-email" id="user-email" />
@@ -34,23 +32,10 @@ function Sign(): JSX.Element {
             </div>
           </form>
         </div>
-
-        <footer className="page-footer">
-          <div className="logo">
-            <a href="main.html" className="logo__link logo__link--light">
-              <span className="logo__letter logo__letter--1">W</span>
-              <span className="logo__letter logo__letter--2">T</span>
-              <span className="logo__letter logo__letter--3">W</span>
-            </a>
-          </div>
-
-          <div className="copyright">
-            <p>© 2019 What to watch Ltd.</p>
-          </div>
-        </footer>
+        <Footer />
       </div>
     </React.Fragment>
   );
 }
 
-export default Sign;
+export default SignMessage;
