+++ src/components/user-block/user-block.css
@@ -1,0 +1,4 @@
+.user-block__avatar img {
+	width: 63px;
+	height: 63px;
+}
