+++ src/const.ts
@@ -1,0 +1,9 @@
+export enum AppRoute {
+	Main = '/',
+	Login = '/login',
+	MyList = '/my-list',
+	Films = '/films',
+	Review = ':id/review',
+	Player = '/player/:id'
+}
+
